// src/libs/mrscraper-browser/firefox/server/types.ts
export type ProxySettings = {
  server: string;
  bypass?: string;
  username?: string;
  password?: string;
};
// ...
export type LaunchOptions = channels.BrowserTypeLaunchOptions & {
  useWebSocket?: boolean;
  proxyOverride?: ProxySettings;
};

export type BrowserContextOptions = channels.BrowserNewContextOptions & {
  proxyOverride?: ProxySettings;
  internalIgnoreHTTPSErrors?: boolean;
};

export type ProtocolLogger = (
  direction: "send" | "receive",
  message: object,
) => void;